feat: work and education sections,

+++ src/lib/utils.ts
@@ -21,7 +21,7 @@ export function readingTime(html: string) {
 }
 
 export function dateRange(startDate: Date, endDate?: Date | string): string {
-  const startMonth = startDate.toLocaleString("default", { month: "short" });
+  const startMonth = startDate.toLocaleString("default", { month: "long" });
   const startYear = startDate.getFullYear().toString();
   let endMonth;
   let endYear;
@@ -31,10 +31,10 @@ export function dateRange(startDate: Date, endDate?: Date | string): string {
       endMonth = "";
       endYear = endDate;
     } else {
-      endMonth = endDate.toLocaleString("default", { month: "short" });
+      endMonth = endDate.toLocaleString("default", { month: "long" });
       endYear = endDate.getFullYear().toString();
     }
   }
 
-  return `${startMonth}${startYear} - ${endMonth}${endYear}`;
+  return `${startMonth} ${startYear} - ${endMonth} ${endYear}`;
 }
